+++ examples/nestjs-server/src/app.module.ts
@@ -7,14 +7,14 @@ import { UsersTestModule } from './users-test/users-test.module';
 // Import the configuration module from the local nestjs-env-getter source
 import { AppConfigModule } from 'nestjs-env-getter';
 import { MongoConnectionService } from './mongo-connection.service';
-import { AppConfigOptionsService } from './app-config-options.service';
 
 @Module({
   imports: [
     // Register the custom AppConfig class using nestjs-env-getter with additional providers
     AppConfigModule.forRoot({
       useClass: AppConfig,
-      providers: [AppConfigOptionsService], // Inject the options service
+      // Inject the options service if necessary
+      // providers: [AppConfigOptionsService],
     }),
 
     // Use AppConfig to provide MongoDB connection string from config file